# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/equippableItems": {
            "get": {
                "tags": [
                    "EquippableItems"
                ],
                "summary": "Get all equippable items",
                "description": "Returns a list of all equippable items",
                "operationId": "getAllEquippableItems",
                "responses": {
                    "200": {
                        "description": "A list of equippable items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "equippableItems list stringify"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "EquippableItems"
                ],
                "summary": "Create a random equippable item",
                "description": "Creates a random equippable item based on the given level and ownerId",
                "operationId": "createRandomItem",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "level",
                                    "ownerId"
                                ],
                                "properties": {
                                    "level": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "ownerId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "rarityId": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "blueprintId": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "ownerId": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "randomFactor": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 0.5
                                        },
                                        "damage": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 10.5
                                        },
                                        "health": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 20.5
                                        },
                                        "getRarity": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Rare"
                                                },
                                                "hexColor": {
                                                    "type": "string",
                                                    "example": "#FF0000"
                                                },
                                                "multiplier": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1.5
                                                },
                                                "levelRequiredToDrop": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "getBlueprint": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Sword"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "A sharp sword"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "weapon"
                                                },
                                                "baseDamage": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "baseHealth": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "requiredLevel": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No available rarity or blueprint for the given level"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/usableItems/getUsableItemsOfUser": {
            "get": {
                "tags": [
                    "UsableItems"
                ],
                "summary": "Ottieni gli oggetti utilizzabili di un giocatore",
                "description": "Recupera gli oggetti utilizzabili per un giocatore specifico dato l'ID del proprietario.",
                "operationId": "getUsableItemsOfUser",
                "parameters": [
                    {
                        "name": "ownerId",
                        "in": "query",
                        "description": "ID del giocatore di cui ottenere gli oggetti utilizzabili.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista degli oggetti utilizzabili trovati per il giocatore.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "type": {
                                                "type": "string"
                                            },
                                            "quantity": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Errore nella richiesta: `ownerId` è richiesto.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "ownerId is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Giocatore non trovato con l'ID fornito.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Player not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/usableItems/createRandomUsableItem": {
            "post": {
                "tags": [
                    "UsableItems"
                ],
                "summary": "Crea un oggetto utilizzabile casuale",
                "description": "Genera un oggetto utilizzabile casuale in base al livello specificato. Se viene fornito ownerId, assegna l'oggetto al giocatore. Altrimenti, l'oggetto non sarà assegnato a nessuno.",
                "operationId": "createRandomUsableItem",
                "requestBody": {
                    "description": "Dati per creare un oggetto utilizzabile casuale",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "level"
                                ],
                                "properties": {
                                    "level": {
                                        "description": "Livello del giocatore per determinare la rarità dell'oggetto",
                                        "type": "integer"
                                    },
                                    "ownerId": {
                                        "description": "ID del giocatore a cui assegnare l'oggetto (opzionale)",
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Oggetto utilizzabile casuale creato con successo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Random usable item created successfully"
                                        },
                                        "usableItem": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "healthRecovery": {
                                                    "type": "integer"
                                                },
                                                "imagePath": {
                                                    "type": "string"
                                                },
                                                "rarityId": {
                                                    "type": "integer"
                                                },
                                                "ownerId": {
                                                    "type": "integer",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errore di validazione: campo richiesto mancante o non valido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The level field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Giocatore non trovato o nessun oggetto utilizzabile disponibile per la rarità specificata.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Player not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}